Improve error types
The trait `FromHex` has a method `from_byte_iter` that is meant to be
implemented by users of the library however there is currently no way
for implementations to control the error type. We can rectify this by
adding an associated error type to the trait.

Renames `parse::Error` to `parse::HexToBytesError` so as to be more
specific and also in preparation for adding bytes-to-hex conversion by
way of iterators soon.

Includes addition of a new error type `HexToArrayError` that is used to
implement `FromHex` for fixed size arrays. Also updates the `hexy`
example to use the new API.