Add a mixed case example/test
This library can parse hex strings of either case. Also we do not care
if a hex string contains a mix of uppercase and lowercase, however we
only output in a single case. This means mixed case strings do not
roundtrip.

Add an example and a unit test that shows parsing of mixed case hex
strings.