Fix formatting
Currently the buffered encoder does not handle padding or precision,
this is because it writes in chunks so cannot.

We would like to support both.

Observe that it is unlikely that anyone will use padding or precision
for a hex string that is larger than one chunk of the buffered
encoder (1024 hex characters) so we can just handle slices shorter than
this.

We elect to pad as a string, this means that users can get various
forms of padding/precision.

Add module level examples to the `display` module demonstrating how
padding/precision can be achieved.

Fix: #23