Rename the library to hex
Currently this crate is used using either `use hex-conservative::Foo` or
using the `package` trick in the dependency specification to shorten it
to `hex`. However the manifest provides a `lib` section for library
renaming, we can use this so that the lib can be imported as `hex`.

With this applied the dependency specification becomes `hex = X.Y.Z` and
the usage becomes `use hex::DisplayHex`.

This can be seen to work by checking the `fuzz` crate.