Improve error types used by `FromHex` trait
Update how error handling is done in the `FromHex` trait. Specifically,
remove the invariant on `HexToBytesIter` that the input hex string is
even. Instead pad with a leading 0 if the input string is uneven. The
benefit of this is that creating the iterator becomes infallible.
Implementors are then required to carry out string length checks in
`from_hex`, which we do for the `Vec` and array imlps.