Simplify `FromHex` and decouple errors
The `ArrayFromHexError` contained `FromBytesError` which had a redundant
`OddLengthStringError` since it implies `InvalidLength`. Decoupling them
was quite difficult due to interaction with the `FromHex` trait. However
the `FromHex` trait had another weirdness: it required implementing
`from_byte_iter` which took already-parsed bytes. That looked fine in
theory however it lead to complications and the only benefit of it was
sparing *one* line of code.

In the ideal world we'd have `FromIterator<T> for Result<[T; N], _>` and
then `FromHex` would naturally be implemented for all things that can be
constructed from an iterator of bytes. However such impl is far away and
we prefer supporting arrays today since they are very common.

Thus this change deletes `from_byte_iter` method and makes `from_hex`
mandatory. It also removes useless bouds on associated type `Error` and
finally decouples the error types.

Notably, this changes error returned from the `FromHex` impl on arrays
to return just `InvalidLength` even if the lenght is odd. The logic
first checks the length so the compiler hopefully sees that the length
is always even and can elide the check in `HexToBytesIter::new` after
inlining.