Generalize `HexToBytesIter`
There was no conversion available from iterator of hex digits to
iterator of decoded bytes. `HexToBytesIter` required having a `&str`
which would be unusable for streaming decoding. This change modifies
`HexToBytesIter` to be generic over internal iterator and adds an opaque
iterator for slices which is used when constructing using `new`. Being
opaque it allows using `array_chunks` in the future.

To enforce that hex digits come in pairs the items returned by the inner
iterator are 2-byte arrays.

Further, this adds `new_unchecked` method to avoid the length check if
it's already known that the length is even.