Make things public
In rust bitcoin we want to be able get at the values in the
`OddLengthStringError` and the `InvalidCharError`.

Add getters to return the error types inner values.

Also, niche users (which `rust-bitcoin` is) want access to the hider
errors (currently only `InvalidLengthError`), so make the `error` module
public.

The `InvalidCharError` is part of the public API because it is returned
is `Iterator::Item` - re-export it from the crate root.