Make fmt_hex_exact honour precision
The `fmt_hex_exact` macro uses `pad_integral` to support adding `0x`
prefix which seems like a reasonable thing to do if one thinks of hashes
as integers. It also seems reasonable to be able to limit the precision
of a hash hex string, `pad_integral` explicitly does not support that.

Add manual handling of the `Foramtter::precision` value.