Bump version to 0.2.1
In preparation for release add a changelog entry and bump the version
number to `v0.2.1`.

Note that since we are pre-1.0 and this release has no breaking changes
we technically could do `v0.2.1` - I get more dopamine by bumping the
minor patch version and it is still, I believe, technically allowed.