Return `[u8; 2]` from `Table::byte_to_hex`
For a given byte, `Table::byte_to_hex` returns the two hex characters in
a `[u8; 2]`, not an `ArrayString<2>`.

This gives callers the freedom to interpret the return value as a `&str`
via `str::from_utf8` or a collection of `char`'s.

`as_hex` now runs in 5x fewer "Estimated Cycles" on
iai-callgrind@0.14.0.