Make InvalidCharError support UTF-8
Currently the `InvalidCharError` error returns a `u8` if an invalid
hex character is found. If that invalid character happened to be a
`char` that encodes as UTF-8 using more than one byte (i.e. not ASCII)
then this means the `u8` returned is not the same as the invalid
character in the input string.

Add support for invalid hex characters that fall outside of a single
ASCII byte.

Because there are so many edge cases, this is a best effort. I'm not
claiming to have gotten this right for all input strings. We default to
just returning the byte.

Currently a 4-byte encoded character that falls on an odd position in
the hex string is not correctly handled but this patch is definitely
an improvement none the less.