Add fmt::Write trait bound to HexWriter
Although not totally necessary, without the trait bound the `HexWriter`
type serves no purpose i.e., if called with a type that does not
implement the trait bound the `HexWriter` does not implement
`io::Write`.

Without this one might wonder what 'writer' this should be. `fmt::Write`
is mentioned already in the docs.

Add an explicit trait bound.