Document and test HexWriter constructor
Thanks to a blanket impl of `Write` for `&mut W` we can pass in a
mutable reference to `HexWriter::new` even though the function signature
suggests it takes ownership of the writer.

Document this as suggested in the Rust API guidelines doc.

While we are at it add a unit test to prove it.

ref: https://rust-lang.github.io/api-guidelines/interoperability.html#c-rw-value