Seal the DisplayHex trait
As part of 1.0 preparation we would like to reduce the API surface as
much as possible.

Seal the `DisplayHex` in the usual manner.

Note we use a macro simial to how we implement `DisplayHex` for various
array sizes. Change the column width so that both macro call sites are
identical.