Remove default impl of hex_reserve_suggestion
The current default impl returns 0, this reduces performance if
implementors forget to implement this method and they know up front how
many bytes are needed. The trait is sealed so we are the only
implementors and it turns out we forgot in one place (well many places
because its a macro).

Remove the default implementation of `hex_reserve_suggestion` and
implement it manually for arrays.