WIP: Depend on hex 1.0.0-alpha.0
We are attempting to do some release trickery in order to stabalize hex
more quickly. It was observed that we only need the `hex` error types in
the public API of downstream bitcoin crates that we want to stabalize.
So we created a 1.0 that was just the parsing logci and error types.

Depend on the new `hex 1.0.0-alpha.0` crate and re-export all its types
and traits in the same spot as they were here.