Introduce crate level decoding functions
We would like to release a bare-bones 1.0 version of this crate that
excludes the `FromHex` trait. So that users can still decode vectors
and arrays without the trait introduced two new decoding functions.

Intentionally do not use `decode` as the name incase we later want to
add a more general `decode` function.