Introduce empty crate
Introduce an empty crate and trimmed down CI.

CI runs on merge into `1.x` branch is based on the idea that development
is done on `master` and we only merge code into `1.x` that has already
been tested there. CI here is basically sanity checks. We run the
`run_task` script with

- stable
- msrv
- docs