Simplify errors
We made a bunch of changes to the error types in an effort to make it
possible to update them without breaking semver - this is a lost cause.

Simplify the error types and just fucking commit to the API.

While we are at it make the `error` module private because it is a code
organization thing only. Re-export all the error types at the crate root.