Swap decoding implementations
We'd like to limit the API in 1.0 to the decoding functions in crate
root and omit the `FromHex` trait to make the API smaller and easier to
maintain. Because the trait won't be in 1.0, the functions should not
call into it.

This moves the implementation of decoding from the trait methods to
free-standing functions and calls those functions inside trait methods.