Rename `decode_{ty}` functions to `decode_to_{ty}`
The functions `decode_vec` and `decode_array` sounded like they are
decoding *from* those objects rather than *to* them. Similar crates that
have such free-standing functions refer to the source in their name but
we refer to destination.

This renames them so that they contain `_to` making it obvious it's the
destination not the source.