Expose the error types as enums directly
The error types in the form of `Outer(pub(crate) Inner)` where `Inner`
needs to be public and accessible anyway have a few problems:

* Understanding two types is harder than understanding just one type
* They are more annoying for the consumers to match on
* They take up space in documentation
* Naming them is difficult

Hex decoding should be in principle easy and all possible error types
are fixed, so there shouldn't be future changes. Adding a common field
can still be done simply by adding it to each variant. This even has the
advantage that newly added field is not lost during some conversions
(e.g. if downstream crate knows the length is correct). If fast access
is needed we can just arrange all structs to place the common fields at
the same offset (e.g. beginning).

So I only see two cases where modifying them would be needed:

* Adding a `PhantomData` field - this is highly unlikely in this crate,
  there's no context. We could have the char type generic in the future
  (`char`, `u8`, `core::ascii::Char`) but that would be stored inside
  `InvalidCharError`, so no need for `PhantomData`.
* Making the error boxed internally - we can also just box the inner
  errors individually and it'll only cost an extra `usize` for
  discriminant making the entire error type two `usize`s. This is
  already smaller than `Vec`, so returning it from `decode_to_vec` is
  non-issue and callers can add another layer of box if needed. It is
  worse when returning short arrays but those seem to be rare.

I think that we can do a lot of layout fiddling and the compiler can do
some too, so I think it'll work fine. And any application that is so
performance-sensitive that 1 `usize` is going to make a difference
should probably not parse hex to begin with.

Thus I believe it's worth just exposing the errors and this commit
implements it.