Add `offset` method to relevant error types
When a larger string is parsed that may consist of multiple parts, some
of them being hex strings, the returned errors should carry the position
relative to the start of the entire string. The design of errors so far
required users to create a new type which would either carry the
position of the hex string or lose some information by converting from
our types. This includes a problem that if a different crate in the tree
enables a feature to find all invalid characters the calling crate has
no way to know about it and has to lose the remaining characters.

This can be solved by providing a method to modify all the positions
stored inside the error type. Such method should be added ASAP so that
all libraries can start using it right away. Thus we want to include
this before 1.0 release.

This commit adds the method, taking into account that it will be used
inside `map_err` quite very often.