Add `newer-rust-version` feature
We've decided that we want to use the `if_rust_version` crate to avoid
compiling and running the same build script multiple times across many
crates. However, currently the additional features are just about
`core::error::Error` and thus irrelevant for `std` users (majority).
Depending on it unconditionally would just add dependency.Making the
crate optional is a natural solution.

This change imlpements it. The specific crate used is considered an
internal detail and this will even go away once the native solution is
in MSRV. Thus we rename the feature to better convey the meaning. It is
also documented that this feature may or may not add dependency and,
since we have a way of removing it, it's also documented what will
happen once the feature becomes irrelevant.