Rename error types to convey their semantics
If a downstream library decodes a hex string into a particular type
using a function in this library the returned error type would "leak"
which type the library used internally. The specific type shouldn't
really concern the callers since it can change anyway. What is relevant
though is whether the length is known at compile time or not.

We might also add more functions in the future decoding to different
types and it'd be better to share error types in such case.

This change renames `HexToBytesError` to `DecodeDynSizedBytesError` and
`HexToArrayError` to `DecodeFixedSizedBytesError` so that types are not
mentioned.