rename DecodeFixedSizedBytesError to DecodeFixedLengthBytesError
Everything that we decode is `Sized`. The phrase "fixed sized" is
grammatically weird, and the term `Sized` has a meaning in Rust which is
not really applicable here (or at least, I argue it is equally
applicable to fixed-sized arrays as it is to fixed-sized objects which
point to resizable backing stores).

Kix does not want to use the term "array" because we may want to extend
to other types which are "fixed size" in the same sense. I think the
phrase "fixed length bytes" better captures this.