rename DynSizedBytes -> VariableLengthBytes
In Rust "dyn sized" is (almost) an oxymoron. As before, the intention is
to refer to whether the *backing store* of the bytes being decoded is
resizable.

However `Sized` does not capture the notion of fixed-length and `dyn` is
a complicated thing that doesn't even appear in this library (except in
the signatures of std::error::Error trait methods).