copy some divergent files from master
Not user-visible (except README.md and CHANGELOG.md), just CI and other
development stuff.

The goal of this commit is that when you diff `1.x` vs `master` you get
only deleted lines, no additions. In fact, there are a couple additions
in CI because we had to edit the list of features to test, but there's
only a couple and they are isolated to the .github/ and contrib/ trees.

I removed the compile-tests from CI because these did not seem to make
sense for our very limited 1.0 API. We should restore them when we add
FromHex etc.