Clarify `error` module documentation
The description said "Error code" but the crate has no error codes in
the meaning of numeric codes and referring to the source code in API
documentation sounds weird. It also repeated the name of the crate which
is redundant.

This rephrases it to say "Error types" and adds a short description
along with pointing the readers to the high-level types.