Add a version notice on top of the README
We have a stable and unstable version of the crate. When people interact
with them it could get confusing which one people use. This is solvable
with `git status` or `cargo tree`/reading of `Cargo.toml` but people new
to the crate may not realize this is happening. While we have
"Stabilization strategy" section it is somewhat long and looks like
"just another section", not something important.

This commit adds a short summary with bold highlight which should draw
attention to the most important aspect: that this is a stable but
API-restricted version.