Create CONTRIBUTING.md and move Githooks to it
The Githooks section in README was a bit noisy because the README is
shown at crates.io which doesn't use git for the individual crates. This
kind of stuff really belongs to CONTRIBUTING but that file didn't exist
so far.

This commit adds the file with some basic instructions and moves the
Githooks section to it.