Remove confusing mention of `rust-bitcoin`
While `hex-conservative` is maintained by the `rust-bitcoin` org, it's
confusing to mention "`rust-bitcoin` policy" in the crate's code because
the policy is not actually org-wide policy. And while the `rust-bitcoin`
*crate* heavily influences the policy of this crate, it doesn't need to
match entirely. This crate could have stricter policy for any reason.

Instead of saying that some derives are `rust-bitcoin` policy, this
commit changes it to say that it's the `hex-conservative` policy.