CI: Use rust-bitcoin-maintainer-tools
Currently the CI script is borked because the MSRV variable is
incorrect meaning the pinning is not working. Because pinning makes me
want to rip my eyes out just have a stab at copying what we did in the
`0.3.x` branch and see if that works.

Copy the `rust.yml` file from the `v0.3.x` branch but change the MSRV
to 1.51.0 as required.